Chore: Renaming modules to snake_case
