+++ PySAMP/text_draw.py
